Create project level notification configuration implementation Hook up non managed pop notification configuration to project level
