+++ platform/platform-impl/src/com/intellij/notification/impl/NotificationsConfigurationImpl.java
@@ -29,22 +29,9 @@ import java.util.Map;
   name = "NotificationConfiguration",
   storages = @Storage("notifications.xml")
 )
-public class NotificationsConfigurationImpl extends NotificationsConfiguration implements PersistentStateComponent<Element>,
-                                                                                          Disposable, ApplicationComponent {
+public class NotificationsConfigurationImpl extends AbstractNotificationsConfigurationImpl implements Disposable, ApplicationComponent {
 
-  private static final Logger LOG = Logger.getInstance(NotificationsConfiguration.class);
-  private static final String SHOW_BALLOONS_ATTRIBUTE = "showBalloons";
-  private static final String SYSTEM_NOTIFICATIONS_ATTRIBUTE = "systemNotifications";
-
-  private static final Comparator<NotificationSettings> NOTIFICATION_SETTINGS_COMPARATOR =
-    (o1, o2) -> o1.getGroupId().compareToIgnoreCase(o2.getGroupId());
-
-  private final Map<String, NotificationSettings> myIdToSettingsMap = new THashMap<>();
-  private final Map<String, String> myToolWindowCapable = new THashMap<>();
   private final MessageBus myMessageBus;
-
-  public boolean SHOW_BALLOONS = true;
-  public boolean SYSTEM_NOTIFICATIONS = true;
 
   public NotificationsConfigurationImpl(@NotNull MessageBus bus) {
     myMessageBus = bus;
@@ -77,33 +64,6 @@ public class NotificationsConfigurationImpl extends NotificationsConfiguration i
 
   public static void remove(String... toRemove) {
     getInstanceImpl().doRemove(toRemove);
-  }
-
-  private synchronized void doRemove(String... toRemove) {
-    for (String groupId : toRemove) {
-      myIdToSettingsMap.remove(groupId);
-      myToolWindowCapable.remove(groupId);
-    }
-  }
-
-  @SuppressWarnings("SynchronizationOnLocalVariableOrMethodParameter")
-  @NotNull
-  public static NotificationSettings getSettings(@NotNull String groupId) {
-    NotificationSettings settings;
-    NotificationsConfigurationImpl impl = getInstanceImpl();
-    synchronized (impl) {
-      settings = impl.myIdToSettingsMap.get(groupId);
-    }
-    return settings == null ? getDefaultSettings(groupId) : settings;
-  }
-
-  @NotNull
-  private static NotificationSettings getDefaultSettings(String groupId) {
-    NotificationGroup group = NotificationGroup.findRegisteredGroup(groupId);
-    if (group != null) {
-      return new NotificationSettings(groupId, group.getDisplayType(), group.isLogByDefault(), false);
-    }
-    return new NotificationSettings(groupId, NotificationDisplayType.BALLOON, true, false);
   }
 
   @Override
@@ -148,19 +108,11 @@ public class NotificationsConfigurationImpl extends NotificationsConfiguration i
     }
   }
 
-  @Override
-  public void changeSettings(String groupDisplayName, NotificationDisplayType displayType, boolean shouldLog, boolean shouldReadAloud) {
-    changeSettings(new NotificationSettings(groupDisplayName, displayType, shouldLog, shouldReadAloud));
-  }
-
-  public synchronized void changeSettings(NotificationSettings settings) {
-    String groupDisplayName = settings.getGroupId();
-    if (settings.equals(getDefaultSettings(groupDisplayName))) {
-      myIdToSettingsMap.remove(groupDisplayName);
-    }
-    else {
-      myIdToSettingsMap.put(groupDisplayName, settings);
-    }
+  @NotNull
+  public static NotificationSettings getSettings(@NotNull String groupId) {
+    NotificationsConfigurationImpl impl = getInstanceImpl();
+    NotificationSettings settings = impl.getSettingsNullable(groupId);
+    return settings == null ? getDefaultSettings(groupId) : settings;
   }
 
   public synchronized boolean isRegistered(@NotNull
@@ -168,50 +120,5 @@ public class NotificationsConfigurationImpl extends NotificationsConfiguration i
     return myIdToSettingsMap.containsKey(id) || NotificationGroup.findRegisteredGroup(id) != null;
   }
 
-  @Override
-  public synchronized Element getState() {
-    Element element = new Element("NotificationsConfiguration");
 
-    NotificationSettings[] sortedNotifications = myIdToSettingsMap.values().toArray(new NotificationSettings[0]);
-    Arrays.sort(sortedNotifications, NOTIFICATION_SETTINGS_COMPARATOR);
-    for (NotificationSettings settings : sortedNotifications) {
-      element.addContent(settings.save());
-    }
-
-    //noinspection NonPrivateFieldAccessedInSynchronizedContext
-    if (!SHOW_BALLOONS) {
-      element.setAttribute(SHOW_BALLOONS_ATTRIBUTE, "false");
-    }
-
-    //noinspection NonPrivateFieldAccessedInSynchronizedContext
-    if (!SYSTEM_NOTIFICATIONS) {
-      element.setAttribute(SYSTEM_NOTIFICATIONS_ATTRIBUTE, "false");
-    }
-
-    return element;
-  }
-
-  @Override
-  public synchronized void loadState(@NotNull final Element state) {
-    myIdToSettingsMap.clear();
-    for (Element child : state.getChildren("notification")) {
-      final NotificationSettings settings = NotificationSettings.load(child);
-      if (settings != null) {
-        final String id = settings.getGroupId();
-        LOG.assertTrue(!myIdToSettingsMap.containsKey(id), String.format("Settings for '%s' already loaded!", id));
-        myIdToSettingsMap.put(id, settings);
-      }
-    }
-    doRemove("Log Only");
-
-    if ("false".equals(state.getAttributeValue(SHOW_BALLOONS_ATTRIBUTE))) {
-      //noinspection NonPrivateFieldAccessedInSynchronizedContext
-      SHOW_BALLOONS = false;
-    }
-
-    if ("false".equals(state.getAttributeValue(SYSTEM_NOTIFICATIONS_ATTRIBUTE))) {
-      //noinspection NonPrivateFieldAccessedInSynchronizedContext
-      SYSTEM_NOTIFICATIONS = false;
-    }
-  }
 }
